# packages/shared-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NH-Media Product API
  version: 1.0.0
  description: Native Workspace-scoped Product API. Large media uses direct signed storage URLs.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}}}
      responses: {'200': {description: Session, content: {application/json: {schema: {$ref: '#/components/schemas/Session'}}}}, '401': {$ref: '#/components/responses/Unauthorized'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
...
  responses:
    Unauthorized: {description: Authentication required}
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties: {username: {type: string}, password: {type: string, format: password}}
    Session:
      type: object
      required: [access_token, token_type]
      properties: {access_token: {type: string}, token_type: {const: Bearer}, workspace_id: {type: string}}
...